Import of release hp2xx-3.4.4_a01.tar.gz 3.4.4a1: 11.2.03
Revived the 'STROKED_FONTS' stubs in chardraw.c by attaching a truetype
font renderer to them. Requires freetype2 and a fixed-width truetype font
such as the LucidaTypewriterRegular.ttf font included with the Sun Java
Runtime Environment (not included here - check licensing conditions).
Fixed two bugs in the scanline renderer that could cause it to fill the gaps
in a complex object. Saving and restoring the pen width when drawing characters
requires the current (parse-time) penwidth to be updated (and not just to
place a command to update the penwidth in the temporary file, which will only
be read at render time) - this is now done by Pen_Width_to_tmpfile().
Both xfig and rgip modes were lacking a handler for the pencolor (DEF_PC)
directive (and would fail on colored input). The rgip mode was still using
wrong penwidths. Added lots of typecasts and typing fixes suggested by
'splint -weak'. Removed the '-m486' compiler switch from my personal Makefile
(the one in sources), as some people try to use it directly instead of adapting
the generic one from the makes directory. Added a 'lint' target for code
checking to same Makefile. Fixed bug that prevented -p with numeric argument
instead of the new A-Z range from working.

+++ sources/clip.c
@@ -25,7 +25,7 @@
 #define ClipSurelyOutside(code1,code2) (code1 & code2)
 #define ClipFullyInside(code1,code2) (! (code1 | code2 ))
 
-static int ClipAreaCode(double x1, double y1,
+static short ClipAreaCode(double x1, double y1,
 			double x2, double y2, double px, double py)
 {
 	short code = CLIP_INSIDE;
